pyproject.toml
[tool.poetry]
name = "ecd-intensive-seminar"
version = "0.1.0"
description = ""
authors = ["Facu Palavecino <facundo6991@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
psycopg2-binary = "2.9.2"
apache-airflow = "2.2.1"
python-decouple = "^3.6"
apache-airflow-providers-postgres = "^5.2.2"
jinja2 = "2.11.0"
jupyter = "^1.0.0"
jupyterlab = "^3.5.1"
prophet = "^1.1.1"
seaborn = "^0.12.1"
scikit-learn = "^1.2.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^6.0.0"
isort = "^5.10.1"
pylint = "^2.15.8"
pre-commit = "^2.20.0"
mypy = "^0.991"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
